backend/requirements.txt
boto3>=1.34.0
flask>=3.0.0
Pillow>=10.0.0
numpy>=1.24.0
ultralytics>=8.0.0
opencv-python>=4.8.0
